Effectful.Hasql

Description

Contains the statically dispatched effect, and relevant functions for a single connection-based DB effect. If you need to take a connection from a pool of connections, you could look into hasql-pool-effectful instead.

This adapts the hasql package.

Documentation

data DB' :: Effect

A single connection DB effect

query

Arguments

:: forall (es :: [Effect]) (a :: Type). DB' :> es

=> Session a
Database session containing one or more statements

-> Eff es (Either QueryError a)

Executes a DB session

>>> import qualified Hasql.Session
>>> import Hasql.TH (singletonStatement)
>>> import Effectful (runEff)

>>> session = Session.statement () [singletonStatement | SELECT 1 :: INT |]
>>> runEff . runDB' connection $ query session
Right 1

runDB'

Arguments

:: forall (es :: [Effect]) (a :: Type). IOE :> es

=> Connection
Will be put in the initial environment. Represents a single DB connection

-> Eff (DB' ': es) a

-> Eff es a

Run a DB' effect with the initial representation/environment: a DB connection.