Effectful.Hasql.Pool

Description

Contains the statically dispatched effect, and relevant functions for a pool of connections DB effect. If you only need to handle a single connection, you could look into hasql-effectful instead.

This adapts the hasql-pool package.

Documentation

data DB :: Effect

Connection pool DB effect

use

Arguments

:: forall (es :: [Effect]) (a :: Type). DB :> es

=> Session a
Session to be executed within the DB environment

-> Eff es (Either UsageError a)

Use a connection from the pool to run a DB session, and returns it to the pool when finished.

runDB

Arguments

:: forall (es :: [Effect]) (a :: Type). IOE :> es

=> Pool
Will be put in the initial environment. Represents a DB pool of connections

-> Eff (DB ': es) a

-> Eff es a

Run a DB effect with the initial representation/environment: a DB pool.